+++ examples/PYCO_WIFI/System/scan_engine.py
@@ -1,0 +1,2 @@
+
+## PYCO_WIFI did not support this example code