+++ examples/PYCO_WIFI_MINI/System/get_started.py
@@ -3,7 +3,11 @@ System - get_started.py.
 
 This example demonstrates how to reboot the device
 
-Copyright (c) 2023-2024 WPC Systems Ltd. All rights reserved.
+For other examples please check:
+    https://github.com/WPC-Systems-Ltd/WPC_Stand-alone_Python_release/tree/main/examples
+
+Copyright (c) 2024 WPC Systems Ltd.
+All rights reserved.
 '''
 
 ## WPC