+++ examples/PYCO_WIFI_MINI/System/scan_engine.py
@@ -1,0 +1,2 @@
+
+## PYCO_WIFI_MINI did not support this example code